refactor(base): rename authToken to authState
