+++ packages/@sanity/base/src/client/wrappedClient.ts
@@ -15,7 +15,7 @@ import sanityClient from '@sanity/client'
 import {generateHelpUrl} from '@sanity/generate-help-url'
 import config from 'config:sanity'
 import configureClient from 'part:@sanity/base/configure-client?'
-import {authToken$} from '../datastores/authToken'
+import {authToken$} from '../datastores/authState'
 
 const fallbackConfig = {projectId: 'UNSPECIFIED', dataset: 'UNSPECIFIED'}
 const apiConfig = {
@@ -73,11 +73,7 @@ export const wrappedClient = {
 
     // Subscribe to auth token stream and configure the client depending on token value
     authToken$.subscribe((token) => {
-      if (token) {
-        newClient.config({...newConfig, token})
-      } else {
-        newClient.config({...newConfig, token: undefined})
-      }
+      newClient.config({...newConfig, token})
     })
 
     return newClient
