+++ packages/@sanity/base/src/datastores/authState/config.ts
