+++ packages/@sanity/base/src/datastores/authState/index.ts
