+++ packages/@sanity/base/src/datastores/authState/storage.ts
