+++ packages/@sanity/base/src/datastores/authState/token.ts
